Parse list type properties of the biomaterial entity type.
If library_strategy or analysis_type are present in the source
data (not derived from NGS files), these properties need to be
parsed as lists (there may be multiple strategies or types for
the biomaterial).
This change parses such values (from the source files) by splitting
the input data text using comma as separator.